* 'dev-1' of https://github.com/LegendOfTeam2/Back-end:
  Fix : 프로필 조회 오류 수정
  Fix: 확인용 커밋
  Fix: SingerPost, MakerPost Id strategy 변경
  Fix: BestSong 로직 수정
  Chore : 프로필 조회 및 수정 로직 보완

+++ src/main/java/com/example/rhythme_backend/dto/responseDto/profile/ProfileResponseDto.java
@@ -11,7 +11,7 @@ public class ProfileResponseDto {
     private String nickname;
     private List<String> hashtag;
     private String introduce;
-    private Integer myPostConunt;
+    private Long myPostConunt;
     private Long follower;
     private Long following;
 }
